+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 
 ## [Unreleased]
 
+## [0.5.0] — 2026–09–17
+
 ### Changed
 
 - The configuration now requires `oxlint` version `1.81.0` or higher.
@@ -85,7 +87,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 
 - Initial config for `oxlint`.
 
-[Unreleased]: https://github.com/firefoxic/oxlint-config/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/firefoxic/oxlint-config/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/firefoxic/oxlint-config/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/firefoxic/oxlint-config/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/firefoxic/oxlint-config/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/firefoxic/oxlint-config/compare/v0.1.3...v0.3.0